Work around Sumo Logic's insane problem with source reconfiguration.
Reconfiguring a single source (adding one, renaming an existing one,
dropping a source, etc.) is a huge, massive problem, apparently.  Once
the sources.json file is updated, the Sumo Logic service needs to not
only be restarted, but also be 'clobbered' in order for any sane search
behaviour to materialize.  "Any" doesn't mean "all" in this context,
because any searches involving _collector or _source fail past a
clobber.

This idiotic behaviour means you have to resort to _hostName (OK)
and _sourceName (retarded) to search your logs.  _sourceName, of
course, has the downside that it's also retarded, and any logs that
have been pulled in twice (i.e. before you reconfigured your sources,
and afterwards, when the collector rescans all existing ones from
the beginning) will show up in searches twice.

To work around *that* problem, we trigger logrotate.  I know, I hate
it, it's not portable, etc.  But this solves the issue for us...
sorry folks.  Blame Sumo.